+++ src/entities/post/api/index.ts
@@ -3,3 +3,6 @@ export { getPosts, searchPosts, getPostsByTag } from "./services"
 
 // Query Keys
 export { postQueryKeys } from "./query-keys"
+
+// Hooks
+export { useGetPostDetail } from "./hooks/use-get-post-detail"
