+++ src/entities/post/model/types.ts
@@ -14,12 +14,6 @@ export interface SearchPostQuery {
   q?: string
 }
 
-export interface CreatePostRequest {
-  title: string
-  body: string
-  userId: number
-}
-
 export interface UpdatePostRequest {
   title?: string
   body?: string
