+++ src/features/post/api/hooks/use-update-post-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { UpdatePostRequest, Post } from "../../../../entities/post/model/types"
-import { updatePost } from "../services"
 import { postQueryKeys } from "../../../../entities/post/api/query-keys"
+import { UpdatePostRequest } from "../../../../entities/post/model/types"
+import { updatePost } from "../services"
 
 // 게시물 수정 Mutation
 export const useUpdatePostMutation = () => {
@@ -9,12 +9,9 @@ export const useUpdatePostMutation = () => {
 
   return useMutation({
     mutationFn: ({ id, data }: { id: number; data: UpdatePostRequest }) => updatePost(id, data),
-    onSuccess: (updatedPost: Post) => {
+    onSuccess: (_, request) => {
       // 단건 캐시만 업데이트
-      queryClient.setQueryData(postQueryKeys.detail(updatedPost.id), updatedPost)
-      queryClient.invalidateQueries({
-        queryKey: postQueryKeys.lists(),
-      })
+      queryClient.setQueryData(postQueryKeys.detail(request.id), request)
     },
   })
 }
